Get User By Id Api

+++ SyncingSyllabi.WebApi/SyncingSyllabi.Data/Dtos/Base/BaseTrackedDto.cs
@@ -7,8 +7,8 @@ namespace SyncingSyllabi.Data.Dtos.Base
     public class BaseTrackedDto : BaseDto
     {
         public DateTime DateCreated { get; set; }
-        public int CreatedBy { get; set; }
+        public Int64 CreatedBy { get; set; }
         public DateTime DateUpdated { get; set; }
-        public int UpdatedBy { get; set; }
+        public Int64 UpdatedBy { get; set; }
     }
 }
